Data-Science/CDB - Cancer Detction in Blood/src/deployment/requirements.txt
fastapi==0.115.8
numpy==1.26.0
pillow==11.1.0
PyYAML==6.0.2
jinja2
python-multipart
tensorflow==2.18.0
uvicorn==0.34.0
